requirements.txt
python-telegram-bot>=20.0
openai>=0.27.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.5.0
openpyxl>=3.0.0
fpdf2>=2.7
yfinance>=0.2.0
forex-python>=1.8.0
alpaca-trade-api>=1.4.0
fastapi>=0.85.1
uvicorn>=0.20.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=4.3.0
celery>=5.2.0
pydantic>=1.8.0
streamlit>=1.30.0
catboost>=1.2
python-dotenv>=1.0.0
ccxt>=3.0.0
requests>=2.28.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.0.0
